Refactor: Move getCheckedMediatorPubKeyRing to TradeValidation
