+++ core/src/main/java/bisq/core/trade/protocol/bisq_v1/tasks/maker/MakerProcessesInputsForDepositTxRequest.java
@@ -48,9 +48,9 @@ import java.util.List;
 
 import lombok.extern.slf4j.Slf4j;
 
-import static bisq.core.trade.validation.DepositTxValidation.getCheckedMediatorPubKeyRing;
 import static bisq.core.trade.validation.TradeValidation.checkDSASignature;
 import static bisq.core.trade.validation.TradeValidation.checkPeersDate;
+import static bisq.core.trade.validation.TradeValidation.getCheckedMediatorPubKeyRing;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 @Slf4j
