+++ core/src/main/java/bisq/core/trade/validation/DepositTxValidation.java
@@ -24,7 +24,6 @@ import bisq.core.dao.burningman.DelayedPayoutTxReceiverService;
 import bisq.core.offer.Offer;
 import bisq.core.provider.fee.FeeService;
 import bisq.core.provider.price.PriceFeedService;
-import bisq.core.support.dispute.mediation.mediator.Mediator;
 import bisq.core.trade.protocol.bisq_v1.messages.InputsForDepositTxRequest;
 import bisq.core.user.User;
 
@@ -98,7 +97,7 @@ public final class DepositTxValidation {
                 takerSignatureKey);
         checkPeersDate(request.getCurrentDate());
         NodeAddress mediatorNodeAddress = request.getMediatorNodeAddress();
-        getCheckedMediatorPubKeyRing(mediatorNodeAddress, user);
+        TradeValidation.getCheckedMediatorPubKeyRing(mediatorNodeAddress, user);
         TradePriceValidation.checkTakersTradePrice(request.getTradePrice(), priceFeedService, offer);
         TradeFeeValidation.checkTakerFee(request.getTakerFeeAsCoin(), request.isCurrencyForTakerFeeBtc(), tradeAmount);
         return request;
@@ -223,15 +222,6 @@ public final class DepositTxValidation {
         checkArgument(rawTransactionInputs.stream().allMatch(tradeWalletService::isP2WH),
                 "rawTransactionInputs must not be malleable");
         return rawTransactionInputs;
-    }
-
-    public static PubKeyRing getCheckedMediatorPubKeyRing(NodeAddress mediatorNodeAddress, User user) {
-        checkNotNull(mediatorNodeAddress, "mediatorNodeAddress must not be null");
-        checkNotNull(user, "user must not be null");
-        Mediator mediator = checkNotNull(user.getAcceptedMediatorByAddress(mediatorNodeAddress),
-                "user.getAcceptedMediatorByAddress(mediatorNodeAddress) must not be null");
-        return checkNotNull(mediator.getPubKeyRing(),
-                "mediator.getPubKeyRing() must not be null");
     }
 
     public static void validatePeersInputs(List<RawTransactionInput> rawTransactionInputs,
